refactor(gui.panels): reorganize gui.panels package
