+++ diode_measurement/gui/panels/__init__.py
@@ -1,0 +1,31 @@
+from .k237 import K237Panel
+from .k595 import K595Panel
+from .k2410 import K2410Panel
+from .k2470 import K2470Panel
+from .k2657 import K2657APanel
+from .k2700 import K2700Panel
+from .k4215 import K4215Panel
+from .k6514 import K6514Panel
+from .k6517b import K6517BPanel
+from .a4284a import A4284APanel
+from .e4980a import E4980APanel
+from .brandbox import BrandBoxPanel
+from .k707b import K707BPanel
+from .k708b import K708BPanel
+
+__all__ = [
+    "K237Panel",
+    "K595Panel",
+    "K2410Panel",
+    "K2470Panel",
+    "K2657APanel",
+    "K2700Panel",
+    "K4215Panel",
+    "K6514Panel",
+    "K6517BPanel",
+    "A4284APanel",
+    "E4980APanel",
+    "BrandBoxPanel",
+    "K707BPanel",
+    "K708BPanel",
+]
